compiler: APM nodes track their own substring
I didn't really want to do this as creating the substrings while parsing is a bit of a faff, and it also means each APM struct is bigger than I would like. Nonetheless, it allows us to provide proper error messages when analysing the APM, so it is good enough for now!