compiler: Remove interpreter and create transpiler
The interpreter was fun, but I'm hoping that this will allow me to make more progress quickly as I try to get the language off the ground. For now, it means we must carry the technical debt and burden of having to compile the C output of the transpiler. Long term, I would hope to ditch C in favour of maybe byte code, LLVM, etc.