tests: Update complex type inference of global enum variable
We are currently passing this test, but actually x is not correctly set to the value of y when the program starts. It just so happens we were defaulting to `foo` anyway as the 0 value. This modified version of the test ensures we're testing that x is correctly initialised to the value of y.